requirements.txt
crewai
crewai[tools]
python-dotenv
requests
torch
transformers
sentencepiece